requirements.txt
kubernetes
requests
pyyaml
